Refactor WebhookLoggingCleanup to fix bugs
